requirements.txt
timm
catalyst==20.09
pytorch-lightning
albumentations==1.1.0
ttach
numpy
tqdm
opencv-python
scipy
matplotlib
einops
addict
gdown
yacs
huggingface_hub
fvcore